+++ persistence/test/account_test.go
@@ -179,7 +179,7 @@ func TestAccountsUpdatedAtHeight(t *testing.T) {
 	require.Equal(t, 1, len(accs))
 
 	// Commit & close the context at height 1
-	require.NoError(t, db.Commit(nil))
+	require.NoError(t, db.Commit(nil, nil))
 	// start a new context at height 2
 	db = NewTestPostgresContext(t, 2)
 
@@ -401,7 +401,7 @@ func TestPoolsUpdatedAtHeight(t *testing.T) {
 	require.Equal(t, 1, len(accs))
 
 	// Commit & close the context at height 1
-	require.NoError(t, db.Commit(nil))
+	require.NoError(t, db.Commit(nil, nil))
 	// start a new context at height 2
 	db = NewTestPostgresContext(t, 2)